requirements.txt
gitpython
langchain
langchain_openai
langchain_community
langchain_groq
langgraph